Added test script for PSRP mock, covered the rest of WSMV that I can
