+++ tests/test_xmls/test_run_cmd_get_config_fail/3-response.xml
@@ -1,0 +1,14 @@
+<s:Envelope xmlns:s="http://www.w3.org/2003/05/soap-envelope" xmlns:a="http://schemas.xmlsoap.org/ws/2004/08/addressing" xmlns:x="http://schemas.xmlsoap.org/ws/2004/09/transfer" xmlns:w="http://schemas.dmtf.org/wbem/wsman/1/wsman.xsd" xmlns:rsp="http://schemas.microsoft.com/wbem/wsman/1/windows/shell" xmlns:p="http://schemas.microsoft.com/wbem/wsman/1/wsman.xsd" xml:lang="en-US">
+<s:Header>
+    <a:Action>http://schemas.microsoft.com/wbem/wsman/1/windows/shell/CommandResponse</a:Action>
+    <a:MessageID>uuid:1C6896B9-E87C-40E5-B9F7-E49154517661</a:MessageID>
+    <p:ActivityId>822B1032-A155-0000-1A3F-2B8255A1D201</p:ActivityId>
+    <a:To>http://schemas.xmlsoap.org/ws/2004/08/addressing/role/anonymous</a:To>
+    <a:RelatesTo>uuid:11111111-1111-1111-1111-111111111111</a:RelatesTo>
+</s:Header>
+<s:Body>
+    <rsp:CommandResponse>
+        <rsp:CommandId>0DA15A6E-0960-43FF-B54A-91A1A9D36131</rsp:CommandId>
+    </rsp:CommandResponse>
+</s:Body>
+</s:Envelope>
