+++ tests/test_xmls/test_run_cmd_get_config_fail/4-response.xml
@@ -1,0 +1,21 @@
+<s:Envelope xmlns:s="http://www.w3.org/2003/05/soap-envelope" xmlns:a="http://schemas.xmlsoap.org/ws/2004/08/addressing" xmlns:w="http://schemas.dmtf.org/wbem/wsman/1/wsman.xsd" xmlns:rsp="http://schemas.microsoft.com/wbem/wsman/1/windows/shell" xmlns:p="http://schemas.microsoft.com/wbem/wsman/1/wsman.xsd" xml:lang="en-US">
+<s:Header>
+    <a:Action>http://schemas.microsoft.com/wbem/wsman/1/windows/shell/ReceiveResponse</a:Action>
+    <a:MessageID>uuid:7F8A4987-6CB9-472C-8A86-FC2964E99C79</a:MessageID>
+    <p:ActivityId>822B1032-A155-0000-203F-2B8255A1D201</p:ActivityId>
+    <a:To>http://schemas.xmlsoap.org/ws/2004/08/addressing/role/anonymous</a:To>
+    <a:RelatesTo>uuid:11111111-1111-1111-1111-111111111111</a:RelatesTo>
+</s:Header>
+<s:Body>
+    <rsp:ReceiveResponse>
+        <rsp:Stream Name="stdout" CommandId="0DA15A6E-0960-43FF-B54A-91A1A9D36131">DQpXaW5kb3dzIElQIENvbmZpZ3VyYXRpb24NCg0K</rsp:Stream>
+        <rsp:Stream Name="stdout" CommandId="0DA15A6E-0960-43FF-B54A-91A1A9D36131">ICAgSG9zdCBOYW1lIC4gLiAuIC4gLiAuIC4gLiAuIC4gLiAuIDogU2VydmVyMjAxNg0KICAgUHJpbWFyeSBEbnMgU3VmZml4ICAuIC4gLiAuIC4gLiAuIDogDQo=</rsp:Stream>
+        <rsp:Stream Name="stdout" CommandId="0DA15A6E-0960-43FF-B54A-91A1A9D36131">ICAgTm9kZSBUeXBlIC4gLiAuIC4gLiAuIC4gLiAuIC4gLiAuIDogSHlicmlkDQogICBJUCBSb3V0aW5nIEVuYWJsZWQuIC4gLiAuIC4gLiAuIC4gOiBObw0KICAgV0lOUyBQcm94eSBFbmFibGVkLiAuIC4gLiAuIC4gLiAuIDogTm8NCg0KRXRoZXJuZXQgYWRhcHRlciBFdGhlcm5ldDoNCg0KICAgQ29ubmVjdGlvbi1zcGVjaWZpYyBETlMgU3VmZml4ICAuIDogDQogICBEZXNjcmlwdGlvbiAuIC4gLiAuIC4gLiAuIC4gLiAuIC4gOiBNaWNyb3NvZnQgSHlwZXItViBOZXR3b3JrIEFkYXB0ZXINCiAgIFBoeXNpY2FsIEFkZHJlc3MuIC4gLiAuIC4gLiAuIC4gLiA6IDAwLTE1LTVELTAxLTY4LTBBDQogICBESENQIEVuYWJsZWQuIC4gLiAuIC4gLiAuIC4gLiAuIC4gOiBZZXMNCiAgIEF1dG9jb25maWd1cmF0aW9uIEVuYWJsZWQgLiAuIC4gLiA6IFllcw0KICAgTGluay1sb2NhbCBJUHY2IEFkZHJlc3MgLiAuIC4gLiAuIDogZmU4MDo6Y2VjOjRlOTM6NDdkOmJlMzElNShQcmVmZXJyZWQpIA0KICAgSVB2NCBBZGRyZXNzLiAuIC4gLiAuIC4gLiAuIC4gLiAuIDogMTkyLjE2OC4xLjYoUHJlZmVycmVkKSANCiAgIFN1Ym5ldCBNYXNrIC4gLiAuIC4gLiAuIC4gLiAuIC4gLiA6IDI1NS4yNTUuMjU1LjANCiAgIExlYXNlIE9idGFpbmVkLiAuIC4gLiAuIC4gLiAuIC4gLiA6IE1vbmRheSwgMjAgTWFyY2ggMjAxNyA2OjM5OjQwIFBNDQogICBMZWFzZSBFeHBpcmVzIC4gLiAuIC4gLiAuIC4gLiAuIC4gOiBNb25kYXksIDI3IE1hcmNoIDIwMTcgNjozOTo0MCBQTQ0KICAgRGVmYXVsdCBHYXRld2F5IC4gLiAuIC4gLiAuIC4gLiAuIDogMTkyLjE2OC4xLjENCiAgIERIQ1AgU2VydmVyIC4gLiAuIC4gLiAuIC4gLiAuIC4gLiA6IDE5Mi4xNjguMS4xDQogICBESENQdjYgSUFJRCAuIC4gLiAuIC4gLiAuIC4gLiAuIC4gOiA1MDMzNzExNw0KICAgREhDUHY2IENsaWVudCBEVUlELiAuIC4gLiAuIC4gLiAuIDogMDAtMDEtMDAtMDEtMUYtRTMtRUQtQzQtMDAtMTUtNUQtMDEtNjgtMEENCiAgIEROUyBTZXJ2ZXJzIC4gLiAuIC4gLiAuIC4gLiAuIC4gLiA6IDE5Mi4xNjguMS4xDQogICBQcmltYXJ5IFdJTlMgU2VydmVyIC4gLiAuIC4gLiAuIC4gOiAxOTIuMTY4LjEuMQ0KICAgTmV0QklPUyBvdmVyIFRjcGlwLiAuIC4gLiAuIC4gLiAuIDogRW5hYmxlZA0KDQpUdW5uZWwgYWRhcHRlciBpc2F0YXAue0U0MDQ0MTE5LTM5QkQtNEU4Ni1CQjMyLTA1RDkwNzkxN0MyMn06DQoNCiAgIE1lZGlhIFN0YXRlIC4gLiAuIC4gLiAuIC4gLiAuIC4gLiA6IE1lZGlhIGRpc2Nvbm5lY3RlZA0KICAgQ29ubmVjdGlvbi1zcGVjaWZpYyBETlMgU3VmZml4ICAuIDogDQogICBEZXNjcmlwdGlvbiAuIC4gLiAuIC4gLiAuIC4gLiAuIC4gOiBNaWNyb3NvZnQgSVNBVEFQIEFkYXB0ZXINCiAgIFBoeXNpY2FsIEFkZHJlc3MuIC4gLiAuIC4gLiAuIC4gLiA6IDAwLTAwLTAwLTAwLTAwLTAwLTAwLUUwDQogICBESENQIEVuYWJsZWQuIC4gLiAuIC4gLiAuIC4gLiAuIC4gOiBObw0KICAgQXV0b2NvbmZpZ3VyYXRpb24gRW5hYmxlZCAuIC4gLiAuIDogWWVzDQoNClR1bm5lbCBhZGFwdGVyIFRlcmVkbyBUdW5uZWxpbmcgUHNldWRvLUludGVyZmFjZToNCg0KICAgQ29ubmVjdGlvbi1zcGVjaWZpYyBETlMgU3VmZml4ICAuIDogDQogICBEZXNjcmlwdGlvbiAuIC4gLiAuIC4gLiAuIC4gLiAuIC4gOiBUZXJlZG8gVHVubmVsaW5nIFBzZXVkby1JbnRlcmZhY2UNCiAgIFBoeXNpY2FsIEFkZHJlc3MuIC4gLiAuIC4gLiAuIC4gLiA6IDAwLTAwLTAwLTAwLTAwLTAwLTAwLUUwDQogICBESENQIEVuYWJsZWQuIC4gLiAuIC4gLiAuIC4gLiAuIC4gOiBObw0KICAgQXV0b2NvbmZpZ3VyYXRpb24gRW5hYmxlZCAuIC4gLiAuIDogWWVzDQogICBJUHY2IEFkZHJlc3MuIC4gLiAuIC4gLiAuIC4gLiAuIC4gOiAyMDAxOjA6NWVmNTo3OWZiOjM4MDQ6MzY3ZDozZjU3OmZlZjkoUHJlZmVycmVkKSANCiAgIExpbmstbG9jYWwgSVB2NiBBZGRyZXNzIC4gLiAuIC4gLiA6IGZlODA6OjM4MDQ6MzY3ZDozZjU3OmZlZjklMihQcmVmZXJyZWQpIA0KICAgRGVmYXVsdCBHYXRld2F5IC4gLiAuIC4gLiAuIC4gLiAuIDogOjoNCiAgIERIQ1B2NiBJQUlEIC4gLiAuIC4gLiAuIC4gLiAuIC4gLiA6IDEzNDIxNzcyOA0KICAgREhDUHY2IENsaWVudCBEVUlELiAuIC4gLiAuIC4gLiAuIDogMDAtMDEtMDAtMDEtMUYtRTMtRUQtQzQtMDAtMTUtNUQtMDEtNjgtMEENCiAgIE5ldEJJT1Mgb3ZlciBUY3BpcC4gLiAuIC4gLiAuIC4gLiA6IERpc2FibGVkDQo=</rsp:Stream>
+        <rsp:Stream Name="stdout" CommandId="0DA15A6E-0960-43FF-B54A-91A1A9D36131" End="true"/>
+        <rsp:Stream Name="stderr" CommandId="0DA15A6E-0960-43FF-B54A-91A1A9D36131" End="true"/>
+        <rsp:CommandState CommandId="0DA15A6E-0960-43FF-B54A-91A1A9D36131" State="http://schemas.microsoft.com/wbem/wsman/1/windows/shell/CommandState/Done">
+        <rsp:ExitCode>0</rsp:ExitCode>
+    </rsp:CommandState>
+</rsp:ReceiveResponse>
+</s:Body>
+</s:Envelope>
