+++ tests/test_xmls/test_run_cmd_get_config_fail/6-response.xml
@@ -1,0 +1,10 @@
+<s:Envelope xmlns:s="http://www.w3.org/2003/05/soap-envelope" xmlns:a="http://schemas.xmlsoap.org/ws/2004/08/addressing" xmlns:w="http://schemas.dmtf.org/wbem/wsman/1/wsman.xsd" xmlns:p="http://schemas.microsoft.com/wbem/wsman/1/wsman.xsd" xml:lang="en-US">
+<s:Header>
+    <a:Action>http://schemas.xmlsoap.org/ws/2004/09/transfer/DeleteResponse</a:Action>
+    <a:MessageID>uuid:4D359545-B7EC-45B8-94FF-C28802F789E8</a:MessageID>
+    <p:ActivityId>822B1032-A155-0000-303F-2B8255A1D201</p:ActivityId>
+    <a:To>http://schemas.xmlsoap.org/ws/2004/08/addressing/role/anonymous</a:To>
+    <a:RelatesTo>uuid:11111111-1111-1111-1111-111111111111</a:RelatesTo>
+</s:Header>
+<s:Body/>
+</s:Envelope>
