+++ tests/test_xmls/test_run_ps_script/1-response.xml
@@ -1,0 +1,70 @@
+<s:Envelope xmlns:s="http://www.w3.org/2003/05/soap-envelope" xmlns:a="http://schemas.xmlsoap.org/ws/2004/08/addressing" xmlns:x="http://schemas.xmlsoap.org/ws/2004/09/transfer" xmlns:w="http://schemas.dmtf.org/wbem/wsman/1/wsman.xsd" xmlns:p="http://schemas.microsoft.com/wbem/wsman/1/wsman.xsd" xml:lang="en-US">
+<s:Header>
+    <a:Action>http://schemas.xmlsoap.org/ws/2004/09/transfer/GetResponse</a:Action>
+    <a:MessageID>uuid:1C16D317-215B-48F7-9B7E-4C11CE1A03A9</a:MessageID>
+    <a:To>http://schemas.xmlsoap.org/ws/2004/08/addressing/role/anonymous</a:To>
+    <a:RelatesTo>uuid:11111111-1111-1111-1111-111111111111</a:RelatesTo>
+</s:Header>
+<s:Body>
+    <cfg:Config xmlns:cfg="http://schemas.microsoft.com/wbem/wsman/1/config">
+    <cfg:MaxEnvelopeSizekb>500</cfg:MaxEnvelopeSizekb>
+    <cfg:MaxTimeoutms>60000</cfg:MaxTimeoutms>
+    <cfg:MaxBatchItems>32000</cfg:MaxBatchItems>
+    <cfg:MaxProviderRequests>4294967295</cfg:MaxProviderRequests>
+    <cfg:Client>
+        <cfg:NetworkDelayms>5000</cfg:NetworkDelayms>
+        <cfg:URLPrefix>wsman</cfg:URLPrefix>
+        <cfg:AllowUnencrypted Source="GPO">true</cfg:AllowUnencrypted>
+        <cfg:Auth>
+            <cfg:Basic Source="GPO">true</cfg:Basic>
+            <cfg:Digest Source="GPO">false</cfg:Digest>
+            <cfg:Kerberos>true</cfg:Kerberos>
+            <cfg:Negotiate>true</cfg:Negotiate>
+            <cfg:Certificate>true</cfg:Certificate>
+            <cfg:CredSSP Source="GPO">true</cfg:CredSSP>
+        </cfg:Auth>
+        <cfg:DefaultPorts>
+            <cfg:HTTP>5985</cfg:HTTP>
+            <cfg:HTTPS>5986</cfg:HTTPS>
+        </cfg:DefaultPorts>
+        <cfg:TrustedHosts>*</cfg:TrustedHosts>
+    </cfg:Client>
+    <cfg:Service>
+        <cfg:RootSDDL>O:NSG:BAD:P(A;;GA;;;BA)(A;;GR;;;IU)S:P(AU;FA;GA;;;WD)(AU;SA;GXGW;;;WD)</cfg:RootSDDL>
+        <cfg:MaxConcurrentOperations>4294967295</cfg:MaxConcurrentOperations>
+        <cfg:MaxConcurrentOperationsPerUser>1500</cfg:MaxConcurrentOperationsPerUser>
+        <cfg:EnumerationTimeoutms>240000</cfg:EnumerationTimeoutms>
+        <cfg:MaxConnections>300</cfg:MaxConnections>
+        <cfg:MaxPacketRetrievalTimeSeconds>120</cfg:MaxPacketRetrievalTimeSeconds>
+        <cfg:AllowUnencrypted Source="GPO">true</cfg:AllowUnencrypted>
+        <cfg:Auth>
+            <cfg:Basic Source="GPO">true</cfg:Basic>
+            <cfg:Kerberos>true</cfg:Kerberos>
+            <cfg:Negotiate>true</cfg:Negotiate>
+            <cfg:Certificate>false</cfg:Certificate>
+            <cfg:CredSSP Source="GPO">true</cfg:CredSSP>
+            <cfg:CbtHardeningLevel>Relaxed</cfg:CbtHardeningLevel>
+        </cfg:Auth>
+        <cfg:DefaultPorts>
+            <cfg:HTTP>5985</cfg:HTTP>
+            <cfg:HTTPS>5986</cfg:HTTPS>
+        </cfg:DefaultPorts>
+        <cfg:IPv4Filter Source="GPO">*</cfg:IPv4Filter>
+        <cfg:IPv6Filter Source="GPO"/>
+        <cfg:EnableCompatibilityHttpListener Source="GPO">false</cfg:EnableCompatibilityHttpListener>
+        <cfg:EnableCompatibilityHttpsListener Source="GPO">false</cfg:EnableCompatibilityHttpsListener>
+        <cfg:CertificateThumbprint>85 e3 17 d6 6a d7 99 6c  bc 9d 12 7a a0 7c b8 842e 46 ae b4</cfg:CertificateThumbprint>
+        <cfg:AllowRemoteAccess Source="GPO">true</cfg:AllowRemoteAccess>
+    </cfg:Service>
+    <cfg:Winrs>
+        <cfg:AllowRemoteShellAccess Source="GPO">true</cfg:AllowRemoteShellAccess>
+        <cfg:IdleTimeout>7200000</cfg:IdleTimeout>
+        <cfg:MaxConcurrentUsers>10</cfg:MaxConcurrentUsers>
+        <cfg:MaxShellRunTime>2147483647</cfg:MaxShellRunTime>
+        <cfg:MaxProcessesPerShell>25</cfg:MaxProcessesPerShell>
+        <cfg:MaxMemoryPerShellMB>1024</cfg:MaxMemoryPerShellMB>
+        <cfg:MaxShellsPerUser>30</cfg:MaxShellsPerUser>
+    </cfg:Winrs>
+</cfg:Config>
+</s:Body>
+</s:Envelope>
