+++ tests/test_xmls/test_run_ps_script/2-request.xml
@@ -1,0 +1,26 @@
+<s:Envelope xmlns:plugin="http://schemas.microsoft.com/wbem/wsman/1/config/PluginConfiguration" xmlns:rsp="http://schemas.microsoft.com/wbem/wsman/1/windows/shell" xmlns:cert="http://schemas.microsoft.com/wbem/wsman/1/config/service/certmapping" xmlns:sub="http://schemas.microsoft.com/wbem/wsman/1/subscription" xmlns:wsen="http://schemas.xmlsoap.org/ws/2004/09/enumeration" xmlns:wsp="http://schemas.xmlsoap.org/ws/2004/09/policy" xmlns:w="http://schemas.dmtf.org/wbem/wsman/1/wsman.xsd" xmlns:cfg="http://schemas.microsoft.com/wbem/wsman/1/config" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:wsdl="http://schemas.xmlsoap.org/wsdl" xmlns:m="http://schemas.microsoft.com/wbem/wsman/1/machineid" xmlns:cim="http://schemas.dmtf.org/wbem/wscim/1/common" xmlns:a="http://schemas.xmlsoap.org/ws/2004/08/addressing" xmlns:wsmanfault="http://schemas.microsoft.com/wbem/wsman/1/wsmanfault" xmlns:wsmid="http://schemas.dmtf.org/wbem/wsman/identify/1/wsmanidentity.xsd" xmlns:p="http://schemas.microsoft.com/wbem/wsman/1/wsman.xsd" xmlns:wse="http://schemas.xmlsoap.org/ws/2004/08/eventing" xmlns:xs="http://www.w3.org/2001/XMLSchema" xmlns:s="http://www.w3.org/2003/05/soap-envelope" xmlns:i="http://schemas.microsoft.com/wbem/wsman/1/cim/interactive.xsd">
+<s:Header>
+    <w:OperationTimeout>PT20S</w:OperationTimeout>
+    <a:MessageID>uuid:11111111-1111-1111-1111-111111111111</a:MessageID>
+    <p:DataLocale s:mustUnderstand="false" xml:lang="en-US"/>
+    <w:ResourceURI s:mustUnderstand="true">http://schemas.microsoft.com/powershell/Microsoft.PowerShell</w:ResourceURI>
+    <a:To>https://windows-host:5986/wsman</a:To>
+    <a:ReplyTo>
+        <a:Address s:mustUnderstand="true">http://schemas.xmlsoap.org/ws/2004/08/addressing/role/anonymous</a:Address>
+    </a:ReplyTo>
+    <w:OptionSet>
+        <w:Option Name="protocolversion">2.3</w:Option>
+    </w:OptionSet>
+    <w:Locale s:mustUnderstand="false" xml:lang="en-US"/>
+    <w:MaxEnvelopeSize>512000</w:MaxEnvelopeSize>
+    <a:Action s:mustUnderstand="true">http://schemas.xmlsoap.org/ws/2004/09/transfer/Create</a:Action>
+    <p:SessionId s:mustUnderstand="false">uuid:11111111-1111-1111-1111-111111111111</p:SessionId>
+</s:Header>
+<s:Body>
+    <rsp:Shell ShellId="11111111-1111-1111-1111-111111111111">
+        <rsp:OutputStreams>stdout</rsp:OutputStreams>
+        <creationXml xmlns="http://schemas.microsoft.com/powershell">AAAAAAAAAAEAAAAAAAAAAAMAAADSAgAAAAIAAQARERERERERERERERERERERAAAAAAAAAAAAAAAAAAAAAO+7vzxPYmogUmVmSWQ9IjI4NjMzMTE1MyI+PE1TPjxWZXJzaW9uIE49IlBTVmVyc2lvbiI+Mi4wPC9WZXJzaW9uPjxWZXJzaW9uIE49InByb3RvY29sdmVyc2lvbiI+Mi4zPC9WZXJzaW9uPjxWZXJzaW9uIE49IlNlcmlhbGl6YXRpb25WZXJzaW9uIj4xLjEuMC4xPC9WZXJzaW9uPjwvTVM+PC9PYmo+AAAAAAAAAAIAAAAAAAAAAAMAAAr3AgAAAAQAAQARERERERERERERERERERERAAAAAAAAAAAAAAAAAAAAAO+7vzxPYmogUmVmSWQ9IjI4NjMzMTE1MyI+PE1TPjxJMzIgTj0iTWluUnVuc3BhY2VzIj4xPC9JMzI+PEkzMiBOPSJNYXhSdW5zcGFjZXMiPjE8L0kzMj48T2JqIE49IlBTVGhyZWFkT3B0aW9ucyIgUmVmSWQ9IjI4NjMzMTE1MyI+PFROIFJlZklkPSIyODYzMzExNTMiPjxUPlN5c3RlbS5NYW5hZ2VtZW50LkF1dG9tYXRpb24uUnVuc3BhY2VzLlBTVGhyZWFkT3B0aW9uczwvVD48VD5TeXN0ZW0uRW51bTwvVD48VD5TeXN0ZW0uVmFsdWVUeXBlPC9UPjxUPlN5c3RlbS5PYmplY3Q8L1Q+PC9UTj48VG9TdHJpbmc+RGVmYXVsdDwvVG9TdHJpbmc+PEkzMj4wPC9JMzI+PC9PYmo+PE9iaiBOPSJBcGFydG1lbnRTdGF0ZSIgUmVmSWQ9IjI4NjMzMTE1MyI+PFROIFJlZklkPSIyODYzMzExNTMiPjxUPlN5c3RlbS5UaHJlYWRpbmcuQXBhcnRtZW50U3RhdGU8L1Q+PFQ+U3lzdGVtLkVudW08L1Q+PFQ+U3lzdGVtLlZhbHVlVHlwZTwvVD48VD5TeXN0ZW0uT2JqZWN0PC9UPjwvVE4+PFRvU3RyaW5nPlVua25vd248L1RvU3RyaW5nPjxJMzI+MjwvSTMyPjwvT2JqPjxPYmogTj0iSG9zdEluZm8iIFJlZklkPSIyODYzMzExNTMiPjxNUz48T2JqIE49Il9ob3N0RGVmYXVsdERhdGEiIFJlZklkPSIyODYzMzExNTMiPjxNUz48T2JqIE49ImRhdGEiIFJlZklkPSIyODYzMzExNTMiPjxUTiBSZWZJZD0iMjg2MzMxMTUzIj48VD5TeXN0ZW0uQ29sbGVjdGlvbnMuSGFzaHRhYmxlPC9UPjxUPlN5c3RlbS5PYmplY3Q8L1Q+PC9UTj48RENUPjxFbj48STMyIE49IktleSI+MDwvSTMyPjxPYmogTj0iVmFsdWUiIFJlZklkPSIyODYzMzExNTMiPjxNUz48UyBOPSJUIj5TeXN0ZW0uQ29uc29sZUNvbG9yPC9TPjxJMzIgTj0iViI+NzwvSTMyPjwvTVM+PC9PYmo+PC9Fbj48RW4+PEkzMiBOPSJLZXkiPjE8L0kzMj48T2JqIE49IlZhbHVlIiBSZWZJZD0iMjg2MzMxMTUzIj48TVM+PFMgTj0iVCI+U3lzdGVtLkNvbnNvbGVDb2xvcjwvUz48STMyIE49IlYiPjk8L0kzMj48L01TPjwvT2JqPjwvRW4+PEVuPjxJMzIgTj0iS2V5Ij4yPC9JMzI+PE9iaiBOPSJWYWx1ZSIgUmVmSWQ9IjI4NjMzMTE1MyI+PE1TPjxTIE49IlQiPlN5c3RlbS5NYW5hZ2VtZW50LkF1dG9tYXRpb24uSG9zdC5Db29yZGluYXRlczwvUz48T2JqIE49IlYiIFJlZklkPSIyODYzMzExNTMiPjxNUz48STMyIE49IngiPjA8L0kzMj48STMyIE49InkiPjQ8L0kzMj48L01TPjwvT2JqPjwvTVM+PC9PYmo+PC9Fbj48RW4+PEkzMiBOPSJLZXkiPjM8L0kzMj48T2JqIE49IlZhbHVlIiBSZWZJZD0iMjg2MzMxMTUzIj48TVM+PFMgTj0iVCI+U3lzdGVtLk1hbmFnZW1lbnQuQXV0b21hdGlvbi5Ib3N0LkNvb3JkaW5hdGVzPC9TPjxPYmogTj0iViIgUmVmSWQ9IjI4NjMzMTE1MyI+PE1TPjxJMzIgTj0ieCI+MDwvSTMyPjxJMzIgTj0ieSI+MDwvSTMyPjwvTVM+PC9PYmo+PC9NUz48L09iaj48L0VuPjxFbj48STMyIE49IktleSI+NDwvSTMyPjxPYmogTj0iVmFsdWUiIFJlZklkPSIyODYzMzExNTMiPjxNUz48UyBOPSJUIj5TeXN0ZW0uSW50MzI8L1M+PEkzMiBOPSJWIj4yNTwvSTMyPjwvTVM+PC9PYmo+PC9Fbj48RW4+PEkzMiBOPSJLZXkiPjU8L0kzMj48T2JqIE49IlZhbHVlIiBSZWZJZD0iMjg2MzMxMTUzIj48TVM+PFMgTj0iVCI+U3lzdGVtLk1hbmFnZW1lbnQuQXV0b21hdGlvbi5Ib3N0LlNpemU8L1M+PE9iaiBOPSJWIiBSZWZJZD0iMjg2MzMxMTUzIj48TVM+PEkzMiBOPSJ3aWR0aCI+MTIwPC9JMzI+PEkzMiBOPSJoZWlnaHQiPjMwMDA8L0kzMj48L01TPjwvT2JqPjwvTVM+PC9PYmo+PC9Fbj48RW4+PEkzMiBOPSJLZXkiPjY8L0kzMj48T2JqIE49IlZhbHVlIiBSZWZJZD0iMjg2MzMxMTUzIj48TVM+PFMgTj0iVCI+U3lzdGVtLk1hbmFnZW1lbnQuQXV0b21hdGlvbi5Ib3N0LlNpemU8L1M+PE9iaiBOPSJWIiBSZWZJZD0iMjg2MzMxMTUzIj48TVM+PEkzMiBOPSJ3aWR0aCI+MTIwPC9JMzI+PEkzMiBOPSJoZWlnaHQiPjc5PC9JMzI+PC9NUz48L09iaj48L01TPjwvT2JqPjwvRW4+PEVuPjxJMzIgTj0iS2V5Ij43PC9JMzI+PE9iaiBOPSJWYWx1ZSIgUmVmSWQ9IjI4NjMzMTE1MyI+PE1TPjxTIE49IlQiPlN5c3RlbS5NYW5hZ2VtZW50LkF1dG9tYXRpb24uSG9zdC5TaXplPC9TPjxPYmogTj0iViIgUmVmSWQ9IjI4NjMzMTE1MyI+PE1TPjxJMzIgTj0id2lkdGgiPjEyMDwvSTMyPjxJMzIgTj0iaGVpZ2h0Ij45ODwvSTMyPjwvTVM+PC9PYmo+PC9NUz48L09iaj48L0VuPjxFbj48STMyIE49IktleSI+ODwvSTMyPjxPYmogTj0iVmFsdWUiIFJlZklkPSIyODYzMzExNTMiPjxNUz48UyBOPSJUIj5TeXN0ZW0uTWFuYWdlbWVudC5BdXRvbWF0aW9uLkhvc3QuU2l6ZTwvUz48T2JqIE49IlYiIFJlZklkPSIyODYzMzExNTMiPjxNUz48STMyIE49IndpZHRoIj4xODE8L0kzMj48STMyIE49ImhlaWdodCI+OTg8L0kzMj48L01TPjwvT2JqPjwvTVM+PC9PYmo+PC9Fbj48RW4+PEkzMiBOPSJLZXkiPjk8L0kzMj48T2JqIE49IlZhbHVlIiBSZWZJZD0iMjg2MzMxMTUzIj48TVM+PFMgTj0iVCI+U3lzdGVtLlN0cmluZzwvUz48UyBOPSJWIj5QeXdpbnJtIFBTUlA8L1M+PC9NUz48L09iaj48L0VuPjwvRENUPjwvT2JqPjwvTVM+PC9PYmo+PEIgTj0iX2lzSG9zdE51bGwiPmZhbHNlPC9CPjxCIE49Il9pc0hvc3RVSU51bGwiPmZhbHNlPC9CPjxCIE49Il9pc0hvc3RSYXdVSU51bGwiPmZhbHNlPC9CPjxCIE49Il91c2VSdW5zcGFjZUhvc3QiPmZhbHNlPC9CPjwvTVM+PC9PYmo+PE5pbCBOPSJBcHBsaWNhdGlvbkFyZ3VtZW50cyI+PC9OaWw+PC9NUz48L09iaj4=</creationXml>
+        <rsp:InputStreams>stdin pr</rsp:InputStreams>
+    </rsp:Shell>
+</s:Body>
+</s:Envelope>
