+++ tests/test_xmls/test_run_ps_script/2-response.xml
@@ -1,0 +1,39 @@
+<s:Envelope xmlns:s="http://www.w3.org/2003/05/soap-envelope" xmlns:a="http://schemas.xmlsoap.org/ws/2004/08/addressing" xmlns:x="http://schemas.xmlsoap.org/ws/2004/09/transfer" xmlns:w="http://schemas.dmtf.org/wbem/wsman/1/wsman.xsd" xmlns:rsp="http://schemas.microsoft.com/wbem/wsman/1/windows/shell" xmlns:p="http://schemas.microsoft.com/wbem/wsman/1/wsman.xsd" xml:lang="en-US">
+<s:Header>
+    <a:Action>http://schemas.xmlsoap.org/ws/2004/09/transfer/CreateResponse</a:Action>
+    <a:MessageID>uuid:1B30579A-F1F1-4ED9-AB4D-5A0D0FDEEA1F</a:MessageID>
+    <a:To>http://schemas.xmlsoap.org/ws/2004/08/addressing/role/anonymous</a:To>
+    <a:RelatesTo>uuid:11111111-1111-1111-1111-111111111111</a:RelatesTo>
+</s:Header>
+<s:Body>
+    <x:ResourceCreated>
+        <a:Address>https://10.13.98.89:5986/wsman</a:Address>
+        <a:ReferenceParameters>
+            <w:ResourceURI>http://schemas.microsoft.com/powershell/Microsoft.PowerShell</w:ResourceURI>
+            <w:SelectorSet>
+                <w:Selector Name="ShellId">35829FF7-BDE7-49B7-9B42-835A8158EB42</w:Selector>
+            </w:SelectorSet>
+        </a:ReferenceParameters>
+    </x:ResourceCreated>
+    <rsp:Shell xmlns:rsp="http://schemas.microsoft.com/wbem/wsman/1/windows/shell">
+    <rsp:ShellId>35829FF7-BDE7-49B7-9B42-835A8158EB42</rsp:ShellId>
+    <rsp:ResourceUri>http://schemas.microsoft.com/powershell/Microsoft.PowerShell</rsp:ResourceUri>
+    <rsp:Owner>XINT\A359018</rsp:Owner>
+    <rsp:ClientIP>10.71.3.229</rsp:ClientIP>
+    <rsp:ProcessId>1284</rsp:ProcessId>
+    <rsp:IdleTimeOut>PT7200.000S</rsp:IdleTimeOut>
+    <rsp:InputStreams>stdin pr</rsp:InputStreams>
+    <rsp:OutputStreams>stdout</rsp:OutputStreams>
+    <rsp:MaxIdleTimeOut>PT2147483.647S</rsp:MaxIdleTimeOut>
+    <rsp:Locale>en-US</rsp:Locale>
+    <rsp:DataLocale>en-US</rsp:DataLocale>
+    <rsp:CompressionMode>NoCompression</rsp:CompressionMode>
+    <rsp:ProfileLoaded>Yes</rsp:ProfileLoaded>
+    <rsp:Encoding>UTF8</rsp:Encoding>
+    <rsp:BufferMode>Block</rsp:BufferMode>
+    <rsp:State>Connected</rsp:State>
+    <rsp:ShellRunTime>P0DT0H0M0S</rsp:ShellRunTime>
+    <rsp:ShellInactivity>P0DT0H0M0S</rsp:ShellInactivity>
+</rsp:Shell>
+</s:Body>
+</s:Envelope>
