+++ tests/test_xmls/test_run_ps_script/6-request.xml
@@ -1,0 +1,24 @@
+<s:Envelope xmlns:plugin="http://schemas.microsoft.com/wbem/wsman/1/config/PluginConfiguration" xmlns:rsp="http://schemas.microsoft.com/wbem/wsman/1/windows/shell" xmlns:cert="http://schemas.microsoft.com/wbem/wsman/1/config/service/certmapping" xmlns:sub="http://schemas.microsoft.com/wbem/wsman/1/subscription" xmlns:wsen="http://schemas.xmlsoap.org/ws/2004/09/enumeration" xmlns:wsp="http://schemas.xmlsoap.org/ws/2004/09/policy" xmlns:w="http://schemas.dmtf.org/wbem/wsman/1/wsman.xsd" xmlns:cfg="http://schemas.microsoft.com/wbem/wsman/1/config" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:wsdl="http://schemas.xmlsoap.org/wsdl" xmlns:m="http://schemas.microsoft.com/wbem/wsman/1/machineid" xmlns:cim="http://schemas.dmtf.org/wbem/wscim/1/common" xmlns:a="http://schemas.xmlsoap.org/ws/2004/08/addressing" xmlns:wsmanfault="http://schemas.microsoft.com/wbem/wsman/1/wsmanfault" xmlns:wsmid="http://schemas.dmtf.org/wbem/wsman/identify/1/wsmanidentity.xsd" xmlns:p="http://schemas.microsoft.com/wbem/wsman/1/wsman.xsd" xmlns:wse="http://schemas.xmlsoap.org/ws/2004/08/eventing" xmlns:xs="http://www.w3.org/2001/XMLSchema" xmlns:s="http://www.w3.org/2003/05/soap-envelope" xmlns:i="http://schemas.microsoft.com/wbem/wsman/1/cim/interactive.xsd">
+<s:Header>
+    <w:OperationTimeout>PT20S</w:OperationTimeout>
+    <w:SelectorSet>
+        <w:Selector Name="ShellId">11111111-1111-1111-1111-111111111111</w:Selector>
+    </w:SelectorSet>
+    <a:MessageID>uuid:11111111-1111-1111-1111-111111111111</a:MessageID>
+    <p:DataLocale s:mustUnderstand="false" xml:lang="en-US"/>
+    <w:ResourceURI s:mustUnderstand="true">http://schemas.microsoft.com/powershell/Microsoft.PowerShell</w:ResourceURI>
+    <a:To>https://windows-host:5986/wsman</a:To>
+    <a:ReplyTo>
+        <a:Address s:mustUnderstand="true">http://schemas.xmlsoap.org/ws/2004/08/addressing/role/anonymous</a:Address>
+    </a:ReplyTo>
+    <w:Locale s:mustUnderstand="false" xml:lang="en-US"/>
+    <w:MaxEnvelopeSize>512000</w:MaxEnvelopeSize>
+    <a:Action s:mustUnderstand="true">http://schemas.microsoft.com/wbem/wsman/1/windows/shell/Receive</a:Action>
+    <p:SessionId s:mustUnderstand="false">uuid:11111111-1111-1111-1111-111111111111</p:SessionId>
+</s:Header>
+<s:Body>
+    <rsp:Receive>
+        <rsp:DesiredStream CommandId="11111111-1111-1111-1111-111111111111">stdout</rsp:DesiredStream>
+    </rsp:Receive>
+</s:Body>
+</s:Envelope>
